Use real mesh geometries for bolometer cameras
Replace the CSG apertures with real mesh geometries, with an
overridable material property. This prepares for the removal of CSG
aperture support for bolometer cameras in Cherab core, and also
fixes a bug where the CSG apertures in the divertor bolometers were
partially obscuring each other, reducing the measured etendue of the
detectors.

Update the sensitivity demo to override the camera material, since we
aren't interested in reflections inside the bolometer camera for
tomography (yet...).