Support both MAST and MAST-U equilibria
MAST-U EFIT data uses different signal names to MAST, and some of the
semantics of the individual data types are slightly different. This
commit provides two new classes to deal with MAST and MAST-U
equilibria respectively. Both produce an EFITEquilibrium object for a
requested time, which can be used by Cherab.